Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 
-int a=20;
-int b=10;
+float a=20;
+float b=10;
 int main() {
     scanf("%d", a);
     printf("%s", b);
